web/requirements.txt
python_fasthtml==0.6.4
google-cloud-storage==2.18.2
uvicorn==0.30.6
